pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "prosthesis_leg"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "candlesdk>=1.3.3",
]

[project.scripts]
prosthesis-wizard = "prosthesis_leg.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
